+++ product-service/src/main/java/com/musinsam/productservice/global/ProductResponseCode.java
@@ -1,4 +1,4 @@
-package com.musinsam.productservice.global.config;
+package com.musinsam.productservice.global;
 
 import com.musinsam.common.response.SuccessCode;
 import lombok.Getter;
