+++ product-service/src/main/java/com/musinsam/productservice/presentation/controller/ProductControllerApiV1.java
@@ -3,13 +3,13 @@ package com.musinsam.productservice.presentation.controller;
 import static com.musinsam.common.user.UserRoleType.ROLE_COMPANY;
 import static com.musinsam.common.user.UserRoleType.ROLE_MASTER;
 import static com.musinsam.common.user.UserRoleType.ROLE_USER;
-import static com.musinsam.productservice.global.config.ProductResponseCode.PRODUCT_CREATE_SUCCESS;
-import static com.musinsam.productservice.global.config.ProductResponseCode.PRODUCT_DELETE_SUCCESS;
-import static com.musinsam.productservice.global.config.ProductResponseCode.PRODUCT_GET_LIST_SUCCESS;
-import static com.musinsam.productservice.global.config.ProductResponseCode.PRODUCT_GET_STOCK_SUCCESS;
-import static com.musinsam.productservice.global.config.ProductResponseCode.PRODUCT_GET_SUCCESS;
-import static com.musinsam.productservice.global.config.ProductResponseCode.PRODUCT_UPDATE_STOCK_SUCCESS;
-import static com.musinsam.productservice.global.config.ProductResponseCode.PRODUCT_UPDATE_SUCCESS;
+import static com.musinsam.productservice.global.ProductResponseCode.PRODUCT_CREATE_SUCCESS;
+import static com.musinsam.productservice.global.ProductResponseCode.PRODUCT_DELETE_SUCCESS;
+import static com.musinsam.productservice.global.ProductResponseCode.PRODUCT_GET_LIST_SUCCESS;
+import static com.musinsam.productservice.global.ProductResponseCode.PRODUCT_GET_STOCK_SUCCESS;
+import static com.musinsam.productservice.global.ProductResponseCode.PRODUCT_GET_SUCCESS;
+import static com.musinsam.productservice.global.ProductResponseCode.PRODUCT_UPDATE_STOCK_SUCCESS;
+import static com.musinsam.productservice.global.ProductResponseCode.PRODUCT_UPDATE_SUCCESS;
 
 import com.musinsam.common.aop.CustomPreAuthorize;
 import com.musinsam.common.resolver.CurrentUser;
